+++ tests/func/Connectivity/AbstractTest.php
