+++ tests/func/Connectivity/CurrenciesTest.php
