+++ tests/func/Connectivity/PaymentMethodsTest.php
