+++ tests/func/Connectivity/ProductsTest.php
