+++ tests/func/Rate/RateLimitsTest.php
