+++ tests/unit/Request/RequestWithErrorManagementTest.php
@@ -6,7 +6,7 @@
  * @link https://github.com/MockingMagician/coinbase-pro-sdk/blob/master/README.md
  */
 
-namespace MockingMagician\CoinbaseProSdk\Tests\Unit;
+namespace MockingMagician\CoinbaseProSdk\Tests\Unit\Request;
 
 use MockingMagician\CoinbaseProSdk\Contracts\Request\RequestInterface;
 use MockingMagician\CoinbaseProSdk\Functional\Error\ApiError;
